got rid of commented out code
